1-byte Checksum and Wait for RTS Before Transmitting Next Packet

+++ include/vexbridge/utils/checksum.hpp
@@ -10,15 +10,15 @@ namespace vexbridge
     struct Checksum
     {
         /**
-         * Calculates the checksum of a byte array using a simple sum algorithm.
+         * Calculates the 8-bit checksum of a byte array using a simple sum algorithm.
          * @param buffer The buffer to calculate the checksum of.
          * @param length The length of the buffer.
          * @return The checksum of the buffer.
          */
-        static uint16_t calc(uint8_t *buffer, uint16_t length)
+        static uint8_t calc(uint8_t *buffer, size_t length)
         {
-            uint16_t checksum = 0;
-            for (uint16_t i = 0; i < length; i++)
+            uint8_t checksum = 0;
+            for (size_t i = 0; i < length; i++)
                 checksum += buffer[i];
             return checksum;
         }
